api_gateway/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-jose[cryptography]
passlib[bcrypt]
httpx
psycopg2-binary
python-multipart